Added Country Selection Feature

+++ core/src/test/java/com/projects/aldajo92/mercadolibreproducts/models/ProductDTO.kt
@@ -3,5 +3,5 @@ package com.projects.aldajo92.mercadolibreproducts.models
 data class ProductDTO(
     val id: String,
     val title: String,
-    val price: Int
+    val price: Double
 )
